+++ apps/kyb-app/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/ui@0.3.19
   - @ballerine/common@0.7.31
   - @ballerine/workflow-browser-sdk@0.5.30